Implement skeleton for basic shrink scenario (#1971)
This patch implements a state machine skeleton for a basic shrink scenario based 
on 'transitions' library. It consists of a new 'ggrebalance' tool, which will be 
a single entry point for shrink, expand, and cluster rebalance functionality, 
and 'shrink.py', which contains the state machine itself with the shrink logic. 

The main purpose of this half-MVP is to evaluate the state machine pattern
suitability. Therefore it implements only a limited set of requirements for the
shrink, which allows you to support basic shrink workflow.