Fix '003_recovery_targets.pl' test
Commit dc788668bb2 added a new test case into the '003_recovery_targets.pl'
test. This test case failed on GPDB. Root cause: the test used direct call of
'pg_ctl' to start a PG instance. The call was without specifying GPDB-specific
params. Therefore, postmaster failed with fatal error "contentid (from -C
option) is not specified or is invalid" before reaching the target point, the
test was interested in.

Fix: use 'start()' method of the 'PostgresNode' class. It handles all required
GPDB-specific params when invoking 'pg_ctl'. But, as the test checks a
particular error condition during recovery, postmaster is terminated soon after
its start. And it breaks the check for the postmaster.pid file in 'start()'. To
handle it, flag 'fail_ok' is propagated into the '_update_pid()' function, and
if it is set, we do not fail if there are problems with postmaster.pid.