Adjust bfv_legacy test (#2052)
Commit 50fc694 in particular changed semantic of the CREATE OR REPLACE LANGUAGE.
Now as stated in comment in gram.y: "OR REPLACE" is silently translated to
"IF NOT EXISTS", which isn't quite the same, but seems more useful than
throwing an error.

So now when language exists, CREATE OR REPLACE LANGUAGE emits a notice.

This commit adjusts the test bfv_legacy to see this notice. Also we now drop
language if it is exists in the beginning of the test, so test is independent of
language was loaded or not in previous tests.