Fix test output in src/test/regress/expected/opr_sanity.out
PostgreSQL commit 612a1ab added the equalimage B-tree support functions and
updated the related test in opr_sanity.out.

The test fails in GPDB because GPDB has a custom type complex (see
src/include/utils/complex_type.h), which is not present in upstream PostgreSQL
and therefore was not included in that commit.

Should we add deduplication support for this GPDB-specific type?

Currently it does not support it. The equalimage function is supposed to
guarantee that values considered equal by the operator class are also bitwise
identical.  I believe we should not add this support: since the complex type
uses floating-point components, it cannot guarantee that logical equality
implies identical byte representation. PostgreSQL’s own float4 and float8 types
are also marked as not equalimage-compatible for the same reason.

Therefore, just update the test result by keeping the first column (proc) empty
for complex_ops, indicating that it does not support btvarstrequalimage.