Fix behavior of src/backend/storage/lmgr/lock.c
Commit 15ef6ff added a new assertion to the LockAcquireExtended function in
src/backend/storage/lmgr/lock.c. This assertion is invalid on the executor when
inserting into an AO table. For example, the choose_segno_internal function
explicitly acquires an ExclusiveLock lock from LockRelationForExtension and
then calls the GetAppendOnlyEntryAuxOids function, which opens the table_open
table with an AccessShareLock lock. This leads to the same LockAcquireExtended
function, but with IsRelationExtensionLockHeld already set, which is precisely
what the new assertion prevents.