Fix conflicts in src/backend/statistics/dependencies.c (#2187)
Commit 87779aa in src/backend/statistics/dependencies.c in the
dependencies_clauselist_selectivity function replaced most of the code with a
call to the new clauselist_apply_dependencies function. However, the earlier
commit 19cd1cf already added false as a new use_damping argument when calling
the clause_selectivity function in this code. Add false as a new use_damping
argument when calling the clauselist_selectivity_simple function in the new
clauselist_apply_dependencies function.