Refactor locks in AO tables (#2234)
Commit 15ef6ff added a new assertion, !IsRelationExtensionLockHeld, to the
LockAcquireExtended function in src/backend/storage/lmgr/lock.c.

For AO tables, the earlier commit 5778f31 had already added the
LockRelationForExtension lock and the UnlockRelationForExtension unlock to the
AppendOnlyRecycleDeadSegments, AppendOnlyTruncateToEOF, LockSegnoForWrite, and
choose_segno_internal functions in
src/backend/access/appendonly/appendonly/compaction.c and
src/backend/access/appendonly/appendonlywriter.c.

The earlier commit dedd407 moved the LockRelationForExtension lock and the
UnlockRelationForExtension unlock from the function
AppendOptimizedRecycleDeadSegments to the AppendOptimizedDropDeadSegments
function.

This lock type is invalid for use in AO tables, as confirmed by the new
assertion. For example, the choose_segno_internal function explicitly acquires
an ExclusiveLock lock from LockRelationForExtension and then calls the
GetAppendOnlyEntryAuxOids function, which opens the table_open table with an
AccessShareLock lock. This leads to the same LockAcquireExtended function, but
with IsRelationExtensionLockHeld already set, which is precisely what the new
assertion prevents.

Change the lock type in AO tables from LockRelationForExtension to
LockDatabaseObject, and the unlock type from UnlockRelationForExtension to
UnlockDatabaseObject.